Move db connect/execute to separate module
